justice-agent/requirements.txt
# Justice Agent - Python Claude Agent SDK Application
# For UK civil law case assistance

# Claude Agent SDK (latest version)
claude-agent-sdk>=0.1.18

# Environment configuration
python-dotenv>=1.0.0

# HTTP client for API integration
httpx>=0.27.0
aiohttp>=3.9.0

# For async operations
anyio>=4.0.0

# Type checking (development)
mypy>=1.8.0

# Testing (development)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
